Support for inner classes of the parent PApplet
In processing, when you create a class in a tab it's actually an inner
class of the parent PApplet class. This was causing some issues in the
instantiation via reflection. Should be good now.

This doesn't generally support inner classes of _other_ classes however,
but that should be a rare case.